Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -5,6 +5,6 @@ int main(){
     scanf("%d",num1);
     scanf("%d",num2);
     scanf("%d",num3);
-    printf("%.2f",(num1+num2+num3)/3);
+    printf("%.2f",(float)(num1+num2+num3)/3);
     return 0;
 }
